gh-153202: Shrink ob_ref_local to uint8_t in struct _object
Retype the free-threaded PyObject fast local refcount field from
uint32_t to uint8_t (only a few bits are ever used). Field order is
left unchanged: a compiled-layout check (offsetof/sizeof, plus a
ctypes ABI model) shows sizeof(struct _object) stays 32 bytes for
both the original order and a descending-alignment reordering, and for
both ob_ref_shared widths (8-byte Py_ssize_t and 4-byte uint32_t).
The freed 3 bytes only become padding ahead of the 8-byte-aligned
trailing ob_type pointer; crossing below 32 bytes would require
repacking other header fields (out of scope).

Not independently compilable/mergeable: 32-bit atomic wrappers and
arithmetic over ob_ref_local are updated in Tasks 2-5; Task 1 lands
only as part of that atomic set.

+++ Include/object.h
@@ -161,7 +161,7 @@ struct _object {
     uint16_t ob_flags;
     PyMutex ob_mutex;           // per-object lock
     uint8_t ob_gc_bits;         // gc-related state
-    uint32_t ob_ref_local;      // local reference count
+    uint8_t ob_ref_local;       // local reference count (only a few bits used)
     Py_ssize_t ob_ref_shared;   // shared (atomic) reference count
     PyTypeObject *ob_type;
 };
